Added first implementation to git
This implementation contains a C++ library for talking to the RCON port of
Frostbite-based game servers.
There are three flavors of the communications module; synchronous,
asynchronous (requires manual polling) and threaded.
There are a couple of test programs, a console-type application, and a
program that batch invokes the "version" command against a list of
servers.

Both the console and the batch query programs contain hardcoded IP:port
names for now (still waiting for a better UI there).

The batch query program will occasionally hit a pure function call so
there is some unidentified problem (race condition?) related to the
threading. I wouldn't trust the Mutext and ThreadBase classes too much if
I were you.