Cleaned up thread handling; added NonCopyable class trait
The ThreadBase class used to be of fire-and-forget type; the parent had no
way of checking whether the child was still running.

It has been superseded by WaitableThreadBase, where the parent is expected
to explicitly wait for the child to exit before the thread-object is
destroyed. There is also support for the parent to ask the child to exit
cleanly; however this is dependent on the child regularly checking the
status. There is no way to force the child to exit.

Added the NonCopyable trait. Some classes in the library are not designed
to be copyable; inheriting from the NonCopyable class prevents accidental
copying.